graph<int> will work now
Not tested with actually making graph with int parameters, but it'll work.
changed graph<T> to use Index internally, instead of int. Index is a class similar to int and functions like an int. It will resolve the ambiguity when using graph<int> because there are many overloaded functions in graph<T>