partial commit during testing graph<int>
master branch was not updated with the latest code, hence test was not successful as the test cases were made w.r.t. the latest code in mind. First saving this code and updating master with pull request #3. Then will perform the final testing.
Will try to keep it a local commit.