Weighted graphs
Tried a lot of things, for making it success. Learnt a lot of things. Hit a lot of road blocks. Saw a lot of functional faults in techinque. (Like defining null and default weight in constructor for static member variables). Technical faults as in trying to set a default parameter of a function to be a non static member function. Final roadblock was the static member variable has to be defined outside the class. But here I have it as a static member variable of a generic class graph<T,U> with two generic classes. Hence defining them would be pain in the ***. I tried writting this:
template<class T>
Weight<bool> graph<T, bool>::zero = Weight<bool>(0);
And when it gave me error, I didn't even bothered to look at the solution for this online. I just depricated the whole approach and will think something else around it.

Just commiting to remember the stuff I did. will have a reverse commit.
BTW this commit doesn't eve compile.