Added LoE graph
List of Edges, well actually Set of Edges, but it was really a last minute change.
As the only place edges are being used is only add and remove operation in two pure vFunctions.
(This is just saved code, have not run once, as test file for testEdge requires change)