Test run successfull
just with a few modifications I was able to run the code.
Surprized by the fact that no error was there in the code of LoEGraph and vertex. These were written on a span of  20 hours. Started in the morning 7 am then office in between (very interesting piece of work in office today) and then written the final lines now at 12:30 am not feeling alot sleepy.
Happy Day :smile: