Vertex Code improved (learnt new thing)
Made it stable. Learnt a new thing today.
private methods and variables are private at class level not the object level.
Hence here able to write the code in a better way. Now the methods are safe.

Test file for edges and vertices is not in very good condition now. But still seems good enough for edge only (as its code is not much changed); will try to make another test file for vertex and one for LoEGraph.