SegmentTree Added
A generic segment tree.
A segment tree is usefull when you have frequent update operations and equally frequent range queries on a collection of objects.
This is a generic implementation of a segment tree. User can define their own function for query, e.g., addition, multiplication, max, min, etc.
The functionality will not be limited on reporting sum of a range, with the provision of defining own functions user can do alot more.