Update LazySegmentTree.h
1. Added comments
2. updated resolve method, removed its dependency on update method.

Update and resolve are critical methods of this class. Any mistakes in them will make the whole tree structure unstable or inconsistant.

Few things to care:
1. as soon as we update node with its true value, lazy should be made zero (both functions are doing this)
2. If we have passed the lazy parameter correctly (to its children), then the true value calculated from lazy will be consistant with the value from combining children values. (both functions took different approach in here)
a. resolve method is calculating it from the lazy parameter of the node itself, as in case of leaf nodes it may not have the children
b. update method is calculating it from its children's values as it will maintain the consistancy of the tree.

#14 done.